Remove useless folders and unreferenced objects
Take vitashell-themes, remove all folders except translations, move
these translations to "l10n", and remoe unreferenced objects. This was
done in order to keep the entire commit history of changes (the last
thing I would want to do is lose it, especially with git which makes
this easy) and also significantly make the size of the repo smaller; it
went from ~40mb to 300kb, which is much easier to merge to VitaShell.

What about the other folders in vitashell-themes? Well, they're kept
alive in the repo, but it's going to stay unmaintained; they're not
deleted outright to avoid wasting the effort which has been done for
them, and in case someone referenced the vitashell-themes repo for a
non-translation.